// src/app/models/posts.ts
/*posts*/ 
export interface IRegisterPosts {
    title: string;
    content: number; 
    publishDate: string;
    published: boolean; 
    idHouse: string;
    author: string;
  }
  
  export interface IPostsEntity {
    title: string;
    content: number; 
    publishDate: string;
    published: boolean; 
    idHouse: string;
    author: string;
  }